requirements.txt
vllm==0.6.6.post1
funasr==1.2.2
paddlepaddle==3.0.0b1
paddleocr==2.9.1
fastapi==0.115.6
openai==1.59.6
argparse==1.4.0
torchaudio==2.5.1
python-multipart==0.0.20
pydantic-settings==2.2.1
pyyaml==6.0.1
